Rename __tests__ directory in WebRuntime
